Add support for local_version on trydevpi.tryton.org
From pip 20.3 the new solver check that the consistency between the version of
the package and the metadata published.
When installing from the CI a local version is appended by setup.py to generate
a unique version to not reuse the cache.
The local version generated from the CI environment is passed as parameter to
trydevpi so it can generate the proper metadata.
The build os local version is simplified to allow tox substitution to replicate
it.

issue9928
review320501002